--- Profile/ProfileInfo/style.css
@@ -1,10 +1,0 @@
-.profile-info {
-  position: relative;
-  width: 100%;
-  height: 20vh;
-  display: flex;
-  flex-direction: column;
-  align-self: center;
-  justify-content: center;
-  align-items: center;
-}
