--- Settings/Navbar/style.css
@@ -1,4 +1,0 @@
-.nav-bar {
-  display: flex;
-  justify-content: space-between;
-}
