+++ components/Profile/components/ProfileInfo.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import './style.css';
 
 const ProfileInfo = () => {
   return (
