+++ data/interfaces.ts
@@ -1,0 +1,9 @@
+export interface IListing {
+  id: number;
+  name: string;
+  category_id: number;
+  price: number;
+  locality: string;
+  description: string;
+  watch_count: number;
+}
