+++ services/getListings.ts
@@ -1,0 +1,6 @@
+import { IListing } from '../data/interfaces';
+
+export function getListings(): IListing[] {
+  const modal = {} as IListing;
+  return [modal];
+}
